News_InfoGraphicDemo/requirements.txt
streamlit==1.24.0
requests==2.25.1
python-dotenv==1.0.0
pillow==9.5.0
pandas==2.0.3
tos==2.5.5
uuid==1.30